+++ CHANGELOG.md
@@ -35,6 +35,10 @@
 - Breaking change: refactor most options to camel case (instead of underscore separation).
 - Add optional beautifier support for REST responses (parses known numbers stored as strings into numbers).
 
+### 2.0.0-beta.5
+
+- Breaking change: refactor spot getAllCoinsInformation to getBalances().
+
 <!-- ## 2.0.1 -->
   <!-- - Introduction of [coinm-client](./src/coinm-client.ts) for COINM Futures. -->
   <!-- - Introduction of [vanilla-client](./src/vanilla-client.ts) for Vanilla Options. -->